plat-vexpress: fix non-debug build
Fixes a number of build errors when 'PLATFORM=vexpress' and 'DEBUG=':
- '<var> may be used uninitialized'
- 'dereferencing type-punned pointer will break strict-aliasing rules'
This one is addressed by adding -fno-strict-aliasing to the cflags for the
problematic files, which is a temporary solution.
- 'inlining failed'

Also, make sure tee_svc_syscall_table is aligned on a 32-bit boundary.
Otherwise, an alignment fault might occur:
  ERR [0x0] TEE-CORE:tee_pager_print_error_abort:101: data-abort at 0x602f16b
  FSR 0x1 PC 0x6007bd4 TTBR0 0x603804A CONTEXIDR 0x1
  CPUID 0x80000001 DBGPCSR 0x0 CPSR 0x80000013 (read from SPSR)
  ERR [0x0] TEE-CORE:tee_pager_handle_abort:164: [TEE_PAGER] alignement fault!  (trap CPU)

Note: on Foundation_v8, I measured a ~3x speedup between debug and non-debug
builds for asymmetric crypto tests.