hikey: enable SHA1 and SHA256 crypto extensions by default on aarch32
This results in a 3.4x speed improvement on a mixed SHA1 and SHA256 test.
I used xtest 4001, modified to run 2 test loops (one for SHA1 and one for
SHA256) on a 32 KiB buffer. The buffer is hashed 10,000 times in each
loop. I measured the total elapsed time for a test run.
aarch64 numbers are given for comparison although crypto extensions are
not supported in this case.

Test case       # runs    min    avg    max  stddev   relative
                          (s)    (s)    (s)     (%)      speed
--------------  ------  -----  -----  -----  ------  ---------
aarch32, no CE      10  18.65  18.86  19.49     1.7       1.00
aarch32, CE         10   5.49   5.60   5.67     1.0       3.37
aarch64             10  22.02  22.14  22.83     1.2       0.85

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>
Reviewed-by: Pascal Brand <pascal.brand@linaro.org>
Reviewed-by: Jens Wiklander <jens.wiklander@linaro.org>