Fix pager
Commit c974c7ddc7d8 ("plat-stm: use generic linker script") breaks
CFG_WITH_PAGER=y because it adds $(platform-cflags) to the C flags for
version.o. This is correct, otherwise the compiler may generate code
that is not suitable for the platform (such as floating-point
instructions when we don't expect them).

However, compiling version.o with $(platform-cflags) brings in
-fdata-sections which has an unfortunate side effect: the object is not
included in the non-pageable code unless we add it explicitely.
This is what this commit does.

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>
Reviewed-by: Pascal Brand <pascal.brand@linaro.org>
Reviewed-by: Jens Wiklander <jens.wiklander@linaro.org>