ECC: decrease the number of temporary variables
ECC is using a lot (80) temporary variables. These variables
are taken from a static pool, each being of the maximum key size
supported in OP-TEE: 4096bits, times 2 to include
wrapping multiplication in temporary computation.

However, max key size of ECC is only 521bits. So even if there
is a lot of required temporary variables, each variable can be
much smaller than 4096bits.

mp_init_split is now introduced, so that several 521bits variables
can be stored in a single temporary variable of size 4096.

Thanks to this patch, the number of temporary variables is back to 50,
and the emulated esram size (QEMU / FVP / HiKey) is back to 200KB.

Signed-off-by: Pascal Brand <pascal.brand@st.com>