arm: update SHA-256 32-bit CE implementation to process multiple blocks
Adjust the 32-bit ARMv8 Crypto Extensions version of the SHA-256
"compress" function to accept multiple blocks of input data.
Rename a couple of files in preparation for the 64-bit implementation
which will follow, and for consistency with SHA-1.

Performances with various buffer sizes were measured on HiKey with
sha-perf. Values are in MiB/s, column 'n' means no acceleration,
'y (before)' is the parent commit's accelerated code, and 'y (after)'
is this commit.

  Size |  CFG_CRYPTO_SHA256_ARM32_CE=?
 (KiB) |     n | y (before) |  y (after)
 ------+-------+------------+-----------
     1 |  17.8 |       31.9 |      36.3
     2 |  22.9 |       52.1 |      67.4
     4 |  26.9 |       78.9 |     117.5
     8 |  29.4 |      105.2 |     188.4
    16 |  30.9 |      125.3 |     268.5
    32 |  31.7 |      139.4 |     341.7
    64 |  32.1 |      147.8 |     401.4
   128 |  32.4 |      152.4 |     438.7
   256 |  32.5 |      154.8 |     460.6
   384 |  32.5 |      155.4 |     467.0

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>
Reviewed-by: Ard Biesheuvel <ard.biesheuvel@linaro.org>
Reviewed-by: Pascal Brand <pascal.brand@linaro.org>