Add 'make check' target to run xtest in QEMU, use it in Travis builds
This is a RFC (request for comments) pull request that I am submitting
as I am about to leave for a 1-week vacation ;-)
I got this finally working today, so I decided to share. Please let me
know if you think it is the good approach, or if we should do things
differently maybe. Here is what this does.

The check target clones a number of Git repositories under
$(outdir)/check and builds a bootable image containing Linux, OP-TEE
and xtest. Then it runs it in QEMU and reports a pass/fail status.

The Travis configuration file is updated to run "make check" each time
it is invoked. Since it is time-consuming (see below), we may want to
check only pull requests and/or branches that match a given pattern,
TBD.

The "make check" step adds about 12 minutes to the Travis run when
ccache is correctly populated (which should be the case most of the
time, because Travis preserves the cache content between builds).
Add this to the time it takes to run the usual build tests, and the
complete sequence takes 20 minutes, approximately.

Some tools that are required for "make check" are unfortunately not
available in the Travis container infrastructure, which is needed to
enable ccache. Since apt-get is not allowed in that environment, we
have to build the tools from sources: bc, Tcl, Expect.

Note: I have chosen to include the build recipes for the various
projects in mk/check.mk, and also to import the Expect script as
mk/qemu-check.mk rather than use the external 'build' project
(https://github.com/OP-TEE/build). I think it may be easier to
maintain that way.

Note 2: "make check" sometimes fails with Expect error
'send: spawn id exp7 not open' as the kernel is booting, and we are
waiting for the shell prompt to come up (or just after that).
I have not yet investigated this issue.

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>