Travis: run xtest in QEMU
.travis.yml is updated to run functional tests in QEMU after the
multi-platform builds. This involves:

- Cloning Git repositories using the Android 'repo' script (the
manifest is at https://github.com/OP-TEE/manifest)
- Building a bootable image containing Linux, OP-TEE and xtest, then
booting the image in QEMU and running the test suite ("make check")

Travis will report success only if all tests pass.

Some required tools are downloaded and compiled during the
before_install step: bc, Tcl and Expect. Indeed, they are not available
in the Travis container-based infrastructure, which we have to use to
enable ccache. Since apt-get is not allowed in that environment, we
have to build from sources. This is usually quite fast (less than a
minute) thanks to ccache.

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>
Reviewed-by: Joakim Bech <joakim.bech@linaro.org>
Reviewed-by: Pascal Brand <pascal.brand@linaro.org>