hikey: increase shared memory size from 1 to 2 MiB
This is needed to run the latest "generic driver" configuration.
When the shared memory pool is 1 MiB, xtest 7633 fails with a TA panic
due to memory allocation error. This commit increases the size of the
shared memory pool so that the test will pass.
Note that UEFI (EDK2) reserves the top 32 MiB of the physical address
space (0x3E000000-0x3FFFFFFF) for OP-TEE, so we still have 14 MiB unused
(0x3E000000-0x3EDFFFFF).

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>
Reviewed-by: Jens Wiklander <jens.wiklander@linaro.org>
Reviewed-by: Pascal Brand <pascal.brand@linaro.org>