arm/pager: move init temp storage of hashes
When a paged OP-TEE binary is copied into secure memory with unpaged
code and data in the correct location, but with init code and hashes of
paged pages starting at the start of the .bss section. If .bss is large
enough init code and hashes will fit entirely in the .bss section and as
long as .bss is unused the data there is safe.

This assumption will not be true any longer if .bss shrinks dramatically
due to reduced size of mpa scratch memory.

With this patch the hashes will be copied to a temporary safe location
right after the init code. This location is the same as the start of the
.text_pageable section so the hashes must be copied to the final
location before the pager is initialized.

Signed-off-by: Jens Wiklander <jens.wiklander@linaro.org>
Reviewed-by: Jerome Forissier <jerome.forissier@linaro.org>
Reviewed-by: Pascal Brand <pascal.brand@linaro.org>