pager: bug fix boot time moving of hashes
Fixes problem with moving hashes of paged pages during boot. Before MMU is
initialize we invalidate all memory used by TEE Core. In the pager case we
had __init_end as end marker but that's not good now that we store the
hashes temporarily at that address. The new end is now __tmp_hashes_end.

Signed-off-by: Jens Wiklander <jens.wiklander@linaro.org>
Tested-by: Jerome Forissier <jerome.forissier@linaro.org> (HiKey)
Tested-by: Jens Wiklander <jens.wiklander@linaro.org> (Juno)
Reviewed-by: Jerome Forissier <jerome.forissier@linaro.org>