Secure Storage: resolve memory leakage issue
This is caused by forgetting to store the backup_version of
meta file in open_meta_file(). The backup_version field is
random and this may generate two versions of meta file in TEE
file folder at commit time.

Unfortunately, we didn't handle this case well in tee_fs_common_rename().
If we have two meta files in TEE file folder, one buffer created
by strdup() will leak. It's hard to handle this case elegantly
in current implementation, so we just put a warning here.

We will propose a new design to address this issue later.

Signed-off-by: SY Chiu <sy.chiu@linaro.org>
Tested-by: SY Chiu <sy.chiu@linaro.org> (QEMU, MT8173 EVB)
Reviewed-by: Joakim Bech <joakim.bech@linaro.org>