Avoid duplication of $(call force,CFG_WITH_VFP,y)
When cryptographic extensions are used, and whatever the platform,
we have to ensure that VFP preservation is enabled too. Therefore it
makes sense to centralize the tests in core/lib/libtomcrypt/sub.mk
instead of having them in the platform-specific configuration files.

Incidentally, this adds a few missing statements to HiKey and Mediatek.

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>
Reviewed-by: Jens Wiklander <jens.wiklander@linaro.org>
Reviewed-by: Pascal Brand <pascal.brand@linaro.org>