core: RPMB (Replay Protected Memory Block) filesystem support
This is an update to the RPMB filesystem implementation so that the
persistent object API may use RPMB rather than the REE filesystem.
This feature is enabled with CFG_RPMB_FS=y.
Note that this implementation requires support from the non-secure side
to actually access the RPMB partition, as there is no eMMC driver here.
Also, the code is currently not compatible with CFG_ENC_FS (file
encryption), which must be set to 'n'. Encryption will be added later.

Signed-off-by: Youssef Esmat <youssef.esmat@microsoft.com>
Signed-off-by: Paul Swan <Paul.Swan@microsoft.com>
[Rebased onto master, Linux driver/tee-supplicant support]
Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>
Reviewed-by: Jens Wiklander <jens.wiklander@linaro.org>
Reviewed-by: Pascal Brand <pascal.brand@linaro.org>