tee_ta_manager.c: update DMSG() calls to display full TA UUID
The debug traces in tee_ta_manager.c do not show the last 8 bytes of
the TA UUIDs, which makes the output confusing sometimes. This commit
adds a static helper function to tee_misc.h: _uuid2str(), which uses
uuid2str() on a static buffer and can conveniently be used in DMSG().
Care must be taken with reentrancy -- in tee_ta_manager.c we have the
TA lock when we print such traces so there is no problem.

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>