Secure Storage: fixed potential error while creating persistent object
Following is the issue description:

If we failed to create meta file while creating a new persistent
object. We will leave a empty tee folder in ree filesystem.

Next time when we want to create tee file when the same object id,
it will fail because the folder is already existed.

The patch solves this issue by checking if meta file is existed,
instead of just checking if the folder is existed.

Signed-off-by: SY Chiu <sy.chiu@linaro.org>
Tested-by: SY Chiu <sy.chiu@linaro.org> (MT8173)
Reviewed-by: Pascal Brand <pascal.brand@linaro.org>