libutee: use CFG_TEE_PANIC_DEBUG setting
When libutee is compiled, the configuration variable
CFG_TEE_PANIC_DEBUG is not exported to the C code as a pre-processor
macro (contrary to TEE core build, it does not happen automatically).
As a result, all the calls to TEE_Panic() that occur in the
GlobalPlatform API wrappers do not call EMSG() when the TA is about
to panic.
This commit fixes this issue by properly defining the C macro when the
configuration variable is enabled.

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>
Reviewed-by: Pascal Brand <pascal.brand@linaro.org>
Reviewed-by: Jens Wiklander <jens.wiklander@linaro.org>