RPMB filesystem encryption
Adds support for CFG_ENC_FS=y when CFG_RPMB_FS=y.

The files stored on an RPMB partition are encrypted with AES in counter
(CTR) mode. Each file has it own 128-bit File Encryption Key (FEK),
randomly chosen when the file is created, and its own 14-byte nonce,
also randomly generated at creation.
The following are added to the FAT entry for a file:
- the FEK, encrypted with the Secure Storage Key (SSK)
- the nonce, stored in clear text
The file path is reduced accordingly, so that a FAT entry is still 256
bytes.
Each 256-byte block in the file is encrypted in counter mode using the
nonce and the 16-bit block number as the initialization vector:
  IV = (nonce || block number)
  Encrypted block = AES-CTR-Encrypt(AES-Decrypt(SSK, encrypted FEK),
                                    IV, clear text)
  Decrypted block = AES-CTR-Decrypt(AES-Decrypt(SSK, encrypter FEK),
                                    IV, cypher text)

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>

Conflicts:
	core/arch/arm/tee/tee_rpmb.c