core: thread: release locked unused stack pages
Before this patch when the pager is active the thread stacks uses demand
allocation to add required physical pages to the stacks, but those pages
are never returned to the pool of physical pages.

With this patch when pager is active the locked but currently unused
thread stack pages are released on thread free or RPC. This returns the
previously demand allocated physical pages to the pager.

Reviewed-by: Jerome Forissier <jerome.forissier@linaro.org>
Tested-by: Jerome Forissier <jerome.forissier@linaro.org> (HiKey)
Reviewed-by: David Brown <david.brown@linaro.org>
Tested-by: David Brown <david.brown@linaro.org> (qemu-32, FVP-64)
Tested-by: Jens Wiklander <jens.wiklander@linaro.org> (QEMU)
Signed-off-by: Jens Wiklander <jens.wiklander@linaro.org>