arm: fix stack size
According to GP Internal API, TA_STACK_SIZE corresponds to the stack
size used by the TA code itself and does not include stack space
possibly used by the Trusted Core Framework.
Hence, stack_size which is the size of the stack to use,
must be enlarged.

Without this patch, on FVP, xtest 1012, based on ta/sims, fails because
TA_STACK_SIZE is defined as 1024, which is too low.

Reviewed-by: Jerome Forissier <jerome.forissier@linaro.org>
Reviewed-by: Jens Wiklander <jens.wiklander@linaro.org>
Signed-off-by: Pascal Brand <pascal.brand@st.com>