mk/compile.mk: add 'remove' flags variable for submodule
With this commit it is possible to remove a CPP/C/AS flag for the
submodule being built. For instance suppose one adds the following
to core/core.mk:
  cflags-remove-core += -pedantic
...then the '-pedantic' flag will be stripped off from the command
line when building any C file of the TEE core.

Previously, one could remove flags only on a global basis or for a
single file.

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>
Reviewed-by: Pascal Brand <pascal.brand@linaro.org>
Reviewed-by: Jens Wiklander <jens.wiklander@linaro.org>