core: send traces to tee-supplicant when CFG_LOG_SYSLOG=y
Enables integration of TEE messages with the normal world system log
facility (such as syslog).

When CFG_LOG_SYSLOG=y (default: n), log messages are not sent to the
console but to tee-supplicant which then forwards them to the normal
world system logger. "Log messages" means anything logged through
DMSG() etc. or passed to syscall_log(). So, both the TEE core and the
TAs can benefit from this mechanism.

Note that logs are redirected to tee-supplicant only when the first
"load TA" request is received. Before this point, the console is used.

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>
Suggested-by: Zoltan Kuscsik <zoltan.kuscsik@linaro.org>