asop: Add common Makefile for core and TA's
With this fix it's sufficient to add the following to the respective
Android.mk files in the TA directories:

  LOCAL_PATH := $(call my-dir)
  local_module := 11111111-2222-3333-4444444444444444.ta
  include $(BUILT_OPTEE_MK)

Where BUILT_OPTEE_MK is the variant pointing to this aosp_optee.mk, which
has been set in the device.mk for a particular platform.

Doing so gives us the ability to control where the TA files will be
installed, where the intermediates will be generated, and other things
in a single place.

The contents contained in this mk file is mainly related to OP-TEE
and do normally not need to be modified, except when the OP-TEE team
wants to change some behavior. This is why we put this mk file here
instead of putting it into the platform device config repository.

Signed-off-by: Yongqin Liu <yongqin.liu@linaro.org>
Reviewed-by: Joakim Bech <joakim.bech@linaro.org>