assert() honor NDEBUG; replace TEE_ASSERT() with panic_unless()
The assert() macro should expand to a no-op when NDEBUG is defined.
Current implementation provides a TEE_ASSERT() that aborts execution
(if condition fails) whatever debug configuration is. For clarity,
TEE_ASSERT() is renamed panic_unless() and declared in panic.h.

This patch removes several inclusions on tee_common_unpg.h as it used
to declare TEE_ASSERT() and now TEE_ASSERT() is declared in panic.h.
This change lead to including few string.h here and there.

Signed-off-by: jerome forissier <jerome.forissier@linaro.org>
Signed-off-by: etienne carriere <etienne.carriere@linaro.org>