Persistent objects : save informations after close
When persistent objects are created, a header is written
into storage with its sizes. However, if additional data
is written into the persistent object, the header is not
updated properly

This fix adds a new pointer to save a copy of the header
in struct tee_pobj when it is opened or created. If
additional data is written during TEE_WriteObjectData call,
it will re-update the header in storage

Signed-off-by: Hoi-Ho Chan <hoiho.chan@gmail.com>