travis: fix coding style path exclusion
Commit 0bbd70a5c13c ("travis: exclude some paths when running the
coding style tool") is broken in several ways:

- It will not properly exclude the specified paths from the diff output
because the exclamation mark (!) is not correctly protected when
setting $_CP_EXCL. Let's use ":(exclude)" instead of ":!".
- Even with the above fixed, Travis runs an older version of Git that
does not support pathspec magic ":!" or ":(exclude)". Address that by
compiling Git 2.9.3 from source. This adds about 60-75 seconds to the
build time with a cold cache, 8 seconds when the cache is warm.

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>
Reviewed-by: Joakim Bech <joakim.bech@linaro.org>