travis: run coding style tool on each commit in a pull request
`git format-patch --stdout -1` do not work as expected. In addition to
checking only one commit (as opposed to all the commits in a PR), it
may pick the wrong commit. For instance, when the PR contains several
commits, it shows the tip of the target branch instead of the tip of
the PR branch. So we have to modify the way the commits are selected.

There are two reasons why a Travis build is triggered: either a pull
request, or a branch is pushed to a monitored repository. One can tell
the two cases appart thanks to the $TRAVIS_PULL_REQUEST environment
variable.

- In the first case (pull request), HEAD is always a merge commit
between the development branch and the target branch. Use `git
rev-list` to expand the list of commits in HEAD^1..HEAD^2 (those that
are in the dev branch but not in the target branch) and ask
format-patch to show exactly those commits with `-1 <commit-sha1>`.
As a result, all the patches will be checked individually.
- In the second case (not a pull request), HEAD is typically not a
merge, it is simply the tip of the development branch being monitored.
Using '-1 HEAD' is fine.

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>
Reviewed-by: Joakim Bech <joakim.bech@linaro.org>